move function to the HomeGallery component
